+++ src/prefab.py
@@ -1,0 +1,2 @@
+# for storing pre fabricated entities (grouped entities)
+